+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2207](https://github.com/open-telemetry/opentelemetry-python/pull/2207))
 - remove `X-B3-ParentSpanId` for B3 propagator as per OpenTelemetry specification
   ([#2237](https://github.com/open-telemetry/opentelemetry-python/pull/2237))
+- Return proxy instruments from ProxyMeter
+  [[#2169](https://github.com/open-telemetry/opentelemetry-python/pull/2169)]
+- Make Measurement a concrete class
+  ([#2153](https://github.com/open-telemetry/opentelemetry-python/pull/2153))
+- Add metrics API
+  ([#1887](https://github.com/open-telemetry/opentelemetry-python/pull/1887))
 
 ## [1.6.2-0.25b2](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.6.2-0.25b2) - 2021-10-19